fix(controls,photos): switch to infractions tab only when clicking on report infractions
